% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.mb.ecors.R
\name{count.mb.ecors}
\alias{count.mb.ecors}
\title{Count pixels of each Land Use class}
\usage{
\method{count}{mb.ecors}(x, save.format = "ods", save.metadata = T, spreadsheet.folder = getwd())
}
\arguments{
\item{x}{mb.ecors object (from get.mb.ecors).}

\item{save.format}{select save results in "ods" or "RData" format (or both). Set to NULL if you do not want to save files.}

\item{save.metadata}{should save metadadata (txt)?}

\item{spreadsheet.folder}{local folder to save files.}
}
\value{
List with result tables and metadata. If enabled, ods (recomended) or RData files will be saved with result tables an metadata.
}
\description{
Generate tables of number of pixels of each Land Use class in each reference polygon (site, samples, polygons - as defined in function get.mb.ecors) in each year.
}
\examples{
#get a mb.ecors class object
FAL.IBGE.JBB<-st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.retangles<-st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

mb2000_2010<-get.mb.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles,
     polygons=NULL, id.column=1, projected=F, custom.crs=32723,
     collection.mb=6, years=c(2000,2010), resolution=30, evaluate="surroundings.samples",
     buffer1=500, buffer2=1000, buffer3=2000, cumulative.surroundings=F)

table2000_2010<-count.mb.ecors(x=mb2000_2010, save.format=c("ods"))
}
