% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.mb.ecors.R
\name{download.mb.ecors}
\alias{download.mb.ecors}
\title{Download MapBiomas images}
\usage{
download.mb.ecors(
  x,
  exp.degree = 0.05,
  images.folder = getwd(),
  clear.prov = F
)
}
\arguments{
\item{x}{mb.ecors object (from get.mb.ecors).}

\item{exp.degree}{number of degrees to expand (and in all directions) the extent of downloaded images (decimal degree format).}

\item{images.folder}{local folder to save images files.}

\item{clear.prov}{delete ALL files from temporary folder (ecors_temp) in your Google Drive account after download?}
}
\value{

}
\description{
Download from Google Earth Engine all MapBiomas images selected by get.mb.ecors function.
}
\examples{
#get a mb.ecors class object
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.retangles<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

mb2000_2010<-get.mb.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles,
     polygons=NULL, id.column=1, projected=F, custom.crs=32723,
     collection.mb=6, years=c(2000,2010), resolution=30, evaluate="surroundings.site",
     buffer1=10000, buffer2=NULL, buffer3=NULL, cumulative.surroundings=F)

#download
download.mb.ecors(x=mb2000_2010, exp.degree=0.05, images.folder=getwd(), clear.prov=F)
}
