% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ecors.R
\name{get.ecors}
\alias{get.ecors}
\title{Get data to ecors}
\usage{
get.ecors(
  site = NULL,
  points = NULL,
  plots = NULL,
  id.column = 1,
  buffer.points = 1,
  buffer.plots = 0,
  projected = FALSE,
  custom.crs = NULL,
  collection,
  start,
  end,
  bands.eval = NULL,
  bands.vis = T,
  indices = c("NDVI", "EVI", "NBR"),
  resolution,
  eval.area = "site",
  pOK = 0.8,
  singlepixel = F,
  c.dist,
  clouds.sentinel = NULL,
  c.prob = NULL,
  cirrus.threshold = NULL,
  NIR.threshold = NULL,
  CDI.threshold = NULL,
  dmax.shadow = NULL,
  seasons = list(s1 = c(), s2 = c(), s3 = c(), s4 = c()),
  group.by = "month",
  composite = NULL,
  online.storage = "drive"
)
}
\arguments{
\item{site}{polygon of study site (sf object).}

\item{points}{sampling points (sf object).}

\item{plots}{sampling plots (sf object).}

\item{id.column}{number of id.column in your points and/or plots objects (need to be the same for both).}

\item{buffer.points}{radius (m) of buffer of points. Need to be > 0.}

\item{buffer.plots}{radius (m) of buffer of plots.}

\item{projected}{are the provided sf objects projected? (scale = m)}

\item{custom.crs}{choose a crs code to project sf objects prior to buffer processing.}

\item{collection}{Google Earth Engine collection name.}

\item{start}{initial date (year-month-day).}

\item{end}{final date (year-month-day).}

\item{bands.eval}{bands needed for analysis (bands needed to calculate indices or visualization will be added automatically).}

\item{bands.vis}{include bands for visualization?}

\item{indices}{select indices to be produced. Available options are "NDVI", "EVI", "NBR".}

\item{resolution}{select pixel size (m) for (most of) quality control procedures. This value will be passed to stats.ecors and download.ecors.}

\item{eval.area}{choose evaluated area for quality control: site or samples.}

\item{pOK}{minimum proportion of pixels aproved in quality control on the evaluated area to use a image. Use NULL if want use all images.}

\item{c.dist}{additional distance (m) around pre-identified clouds that will be disregarded in the mask (when clouds.sentinel="CDI", c.dist should be multiples of 100 m). Usefull to avoid false negative cloud detection in cloud edges.}

\item{clouds.sentinel}{method for cloud detection in Sentinel-2 MSI. Options are the collection "default" (using built in Quality Analysis of these collections; several false negative pixels; not detect cloud shadows in TOA images), "CDI" (more time consuming) or NULL.}

\item{c.prob}{set the cloud probability threshold value to exclude pixels in Sentinel-2 MSI imagens. It could remove additional pixels with clouds.sentinel="default" (not compatible with "CDI" method).}

\item{cirrus.threshold}{set threshold value for method CDI (see Frantz et al. 2018 for details).}

\item{NIR.threshold}{set threshold value for method CDI (see Frantz et al. 2018 for details).}

\item{CDI.threshold}{set threshold value for method CDI (see Frantz et al. 2018 for details).}

\item{dmax.shadow}{set the maximum distance that CDI Algorithm will search for clouds. Large values take a lot of processing time or may crash Google Earth Engine. Value in meters (in multiples of 100 m).}

\item{seasons}{month allocation (numerical form) in up to four seasons. Use the list structure list(s1=c(), s2=c(), s3=c(), s4=c()) keeping the items you don't want to use empty. More information in the exemples section.}

\item{group.by}{should data be grouped by "season" or "month"? This parameter influences how the data will be summarized in stats.ecors and (if selected) how the images will be composited.}

\item{composite}{method for generating composite images in the collection Available options to argument composite: min, max, mean, median and NULL (disable composition).}

\item{online.storage}{select online storage integration (mandatory for images download). Options are "drive" for Google Drive, "gcs" for Google Cloud Storage or NULL.}
}
\value{
Object of the "ecors" class with metadata and pre-processed data to be used in the stats.ecors, plot.ecors or download.ecors functions. Aditional Google Earth Engine containers objects are exported to .GlobalEnv to be used in rgee functions and avoid errors (elapsed time limit): \cr
\itemize{
\item colle (all images available in the period),
\item colle.filt (images approved in get.ecors quality control),
\item colle.mask (same as the previous one but with bad pixels masked),
\item colle.mask.compo (collection of compositions performed on the images of the previous collection).}
}
\description{
Get data from Google Earth Engine for your study site and integrate with study polygons and sampling periods.
}
\details{
Currently ecors supports the following remote sensing data collections:

#' \cr
Landsat 9
\itemize{
\item "LANDSAT/LC09/C02/T1_L2" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC09_C02_T1_L2}{Collection 2 - Surface Reflectance}
\item "LANDSAT/LC09/C02/T1_TOA" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC09_C02_T1_TOA}{Collection 2 - Top of Atmosphere Reflectance}
\item "LANDSAT/LC09/C02/T1" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC09_C02_T1}{Collection 2 - Raw Images}
}

\cr
Landsat 8
\itemize{
\item "LANDSAT/LC08/C02/T1_L2" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC08_C02_T1_L2}{Collection 2 - Surface Reflectance}
\item "LANDSAT/LC08/C02/T1_TOA" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC08_C02_T1_TOA}{Collection 2 - Top of Atmosphere Reflectance}
\item "LANDSAT/LC08/C02/T1" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LC08_C02_T1}{Collection 2 - Raw Images}
}

\cr
Landsat 7
\itemize{
\item "LANDSAT/LE07/C02/T1_L2" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LE07_C02_T1_L2}{Collection 2 - Surface Reflectance}
\item "LANDSAT/LE07/C02/T1_TOA" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LE07_C02_T1_TOA}{Collection 2 - Top of Atmosphere Reflectance}
\item "LANDSAT/LE07/C02/T1" \href{https://developers.google.com/earth-engine/datasets/catalog/LANDSAT_LE07_C02_T1}{Collection 2 - Raw Images}
}

\cr
Sentinel-2 MSI (Multispectral Instrument)
\itemize{
\item "COPERNICUS/S2_SR" \href{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2_SR}{Surface Reflectance}
\item "COPERNICUS/S2" \href{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2}{Top of Atmosphere Reflectance}
}

\cr
Global Precipitation Measurement (GPM)
\itemize{
\item "NASA/GPM_L3/IMERG_V06" \href{https://developers.google.com/earth-engine/datasets/catalog/NASA_GPM_L3_IMERG_V06}{Global Precipitation Measurement (GPM) v6 - every three hours}
\item "NASA/GPM_L3/IMERG_MONTHLY_V06" \href{https://developers.google.com/earth-engine/datasets/catalog/NASA_GPM_L3_IMERG_MONTHLY_V06}{Monthly Global Precipitation Measurement (GPM) v6}
}
}
\examples{
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.plots<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

#library(ecors)

# Get data (projecting to UTM 32S zone to performe buffer operations)
d2020<-get.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.plots, buffer.points=500, buffer.plots=500,
    eval.area="site", projected=F, custom.crs=32723,
    collection="LANDSAT/LC08/C02/T1_L2", start=c("2020-01-01"), end=c("2020-12-31"),
    bands.eval="SR_B3", bands.vis=T, indices=c("NDVI"), resolution=30,
    pOK=0.3, c.prob=NULL, c.dist=100, clouds.sentinel=NULL, cirrus.threshold=NULL, NIR.threshold=NULL, CDI.threshold=NULL, dmax.shadow=NULL,
    seasons=list(s1=c(11,12,1,2), s2=c(3,4), s3=c(5,6,7,8), s4=c(9,10)), group.by="month", composite="mean")

d.sentinel<-get.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.plots, buffer.points=500, buffer.plots=500,
        eval.area="samples", projected=F, custom.crs=32723,
        collection="COPERNICUS/S2_SR", start=c("2020-03-01"), end=c("2020-03-10"),
        bands.eval=NULL, bands.vis=T, indices=NULL, resolution=10,
        pOK=0, c.prob=NULL, c.dist=100, clouds.sentinel="CDI", cirrus.threshold=0.01, NIR.threshold=0.15, CDI.threshold=-0.5, dmax.shadow=300,
        seasons=list(s1=c(), s2=c(), s3=c(), s4=c()), group.by="month", composite=NULL)

}
\references{
Pixel quality control information: \cr
Frantz, D., Hass, E., Uhl, A., Stoffels, J., & Hill, J. (2018). Improvement of the Fmask algorithm for Sentinel-2 images: Separating clouds from bright surfaces based on parallax effects. Remote sensing of environment, 215, 471-481. \cr
\url{https://www.usgs.gov/media/files/landsat-8-9-olitirs-collection-2-level-1-data-format-control-book} \cr
\url{https://www.usgs.gov/media/files/landsat-8-9-olitirs-collection-2-level-2-data-format-control-book} \cr
\url{https://www.usgs.gov/media/files/landsat-7-etm-collection-2-level-2-data-format-control-book} \cr
\url{https://www.usgs.gov/media/files/landsat-4-5-tm-collection-2-level-2-data-format-control-book} \cr
\url{https://www.usgs.gov/media/files/landsat-8-collection-1-land-surface-reflectance-code-product-guide} \cr
\url{https://www.usgs.gov/media/files/landsat-4-7-collection-1-surface-reflectance-code-ledaps-product-guide} \cr
\url{https://www.usgs.gov/core-science-systems/nli/landsat/landsat-collection-1-level-1-quality-assessment-band?qt-science_support_page_related_con=0#qt-science_support_page_related_con} \cr
}
