% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.focal.mb.ecors.R
\name{plot.focal.mb.ecors}
\alias{plot.focal.mb.ecors}
\title{Plot mb.ecors object}
\usage{
\method{plot}{focal.mb.ecors}(x, window.radius, zoom = 12, legend = T)
}
\arguments{
\item{x}{mb.ecors object (from get.mb.ecors).}

\item{window.radius}{radius of moving window (m).}

\item{zoom}{initial zoom value.}

\item{legend}{choose to show the legend or not.}
}
\value{
Google Earth Engine containers objects are exported to .GlobalEnv to be used in rgee functions and avoid errors (elapsed time limit):
\itemize{
\item map.samples: sample polygons (evaluation carried out in the surroundings of these polygons),
\item map.site: site polygon (evaluation carried out in the surroundings of this polygon),
\item map.polygons: polygon for land use evaluation (evaluation carried out inside these polygons),
\item map.buffer(i): polygons of buffer areas (if selected in get.mb.ecors);
\item map.mb(i): plot object of a raster image (one object per image),
\item leg.mb: legend.}
}
\description{
Plot MapBiomas images processed by focal (morphological reducer) along with site and sample polygons.
}
\details{
Plot to check the effect of "focal" preprocessing on the removal of isolated pixels and eventual removal of real data (Linear features such as gallery forests or water bodies are particularly susceptible to be removed by this technique).
Moving window size indicates area in which values will be replaced by the "mode" value in the area (larger values cause more aggressive removal/simplification).
}
\examples{
#Get a mb.ecors class object
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.retangles<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

# Get data (projecting to UTM 32S zone to performe buffer operations)
mb2000_2010<-get.mb.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles,
     polygons=NULL, id.column=1, projected=F, custom.crs=32723,
     collection.mb=6, years=c(2000,2010), resolution=30, evaluate="surroundings.site",
     buffer1=5000, buffer2=10000, buffer3=NULL, cumulative.surroundings=F)

#Plotting
plot.focal.mb.ecors(x=mb2000_2010,window.radius=80)
}
