% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lu.ecors.R
\name{plot.lu.ecors}
\alias{plot.lu.ecors}
\title{Plot lu.ecors object}
\usage{
\method{plot}{lu.ecors}(x, comp = NULL, zoom = 12, legend = T)
}
\arguments{
\item{x}{lu.ecors object (from get.lu.ecors).}

\item{comp}{lu.ecors object generated by focal.lu.ecors or convert.lu ecors functions or ecors object (remote sensing) used for comparison (optional).}

\item{zoom}{initial zoom value.}

\item{legend}{choose to show the legend or not.}
}
\value{
Google Earth Engine containers objects are exported to .GlobalEnv to be used in rgee functions and avoid errors (elapsed time limit):
\itemize{
\item map.samples: sample polygons (evaluation carried out in the surroundings of these polygons),
\item map.site: site polygon (evaluation carried out in the surroundings of this polygon),
\item map.polygons: polygon for land use evaluation (evaluation carried out inside these polygons),
\item map.buffer(i): polygons of buffer areas (if selected in get.lu.ecors);
\item map.lu(i): plot object of a raster image (one object per image),
\item leg.lu: legend.}
}
\description{
Plot MapBiomas images processed by get.lu.ecors along with site and sample polygons.
}
\examples{
# Get a lu.ecors class object
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.plots<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

# Get data (projecting to UTM 32S zone to performe buffer operations)
lu2000_2010<-get.lu.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.plots,
     polygons=NULL, id.column=1, projected=F, custom.crs=32723,
     collection.lu="mapbiomas9", years=c(2000,2010), resolution=30, evaluate="surroundings.site",
     buffer1=5000, buffer2=10000, buffer3=15000, cumulative.surroundings=F)

# Plotting
plot(x=lu2000_2010)

# Focal post-processing and visually checking
f_lu2000_2010<-focal.lu.ecors(x=lu2000_2010,window.radius=80)
plot(x=lu2000_2010, comp=f_lu2000_2010)

# Grouping several native land use classes and visually checking
converted<-convert.lu.ecors(x=lu2000_2010,old.value = c(1,3,4,10,11,12,13),new.value = rep(1,times=7), new.lu.class = rep("Native",times=7))
plot(x=lu2000_2010, comp=converted)
}
