% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mb.ecors.R
\name{plot.mb.ecors}
\alias{plot.mb.ecors}
\title{Plot mb.ecors object}
\usage{
\method{plot}{mb.ecors}(x, zoom = 12, legend = T)
}
\arguments{
\item{x}{mb.ecors object (from get.mb.ecors).}

\item{zoom}{initial zoom value.}

\item{legend}{choose to show the legend or not.}
}
\value{
Google Earth Engine containers objects are exported to .GlobalEnv to be used in rgee functions and avoid errors (elapsed time limit):
\itemize{
\item map.samples: sample polygons (evaluation carried out in the surroundings of these polygons),
\item map.site: site polygon (evaluation carried out in the surroundings of this polygon),
\item map.polygons: polygon for land use evaluation (evaluation carried out inside these polygons),
\item map.buffer(i): polygons of buffer areas (if selected in get.mb.ecors);
\item map.mb(i): plot object of a raster image (one object per image),
\item leg.mb: legend.}
}
\description{
Plot MapBiomas images processed by get.mb.ecors along with site and sample polygons.
}
\examples{
#Get a mb.ecors class object
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))
test.retangles<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))

# Get data (projecting to UTM 32S zone to performe buffer operations)
mb2000_2010<-get.mb.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles,
     polygons=NULL, id.column=1, projected=F, custom.crs=32723,
     collection.mb=6, years=c(2000,2010), resolution=30, evaluate="surroundings.site",
     buffer1=5000, buffer2=10000, buffer3=NULL, cumulative.surroundings=F)

#Plotting
plot.mb.ecors(x=mb2000_2010)
}
