% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakeup.gee.R
\name{wakeup.gee}
\alias{wakeup.gee}
\title{Wakeup Google Earth Engine}
\usage{
wakeup.gee(online.storage = "drive")
}
\arguments{
\item{online.storage}{select online storage integration (mandatory for images download). Options are "drive" for Google Drive, "gcs" for Google Cloud Storage or NULL.}
}
\value{

}
\description{
If you receive an error message regarding the Google Earth Engine server being disconnected due to inactivity, use this command to re-establish the connection.
}
\examples{
wakeup.gee()

}
