% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecors-package.R
\docType{package}
\name{ecors}
\alias{ecors}
\title{ecors: Remote Sensing data acquisition and processing for ecology with Google Earth Engine}
\description{
It simplifies the acquisition of data in Google Earth Engine and
performs several steps of organization and processing typical of ecological studies,
allowing the visualization of maps, data download and production of simple statistics
commonly used in ecological studies. It aims to simplify the use of commands from the
rgee package and the Google Earth Engine platform with a style of functions more
familiar to R users who usually perform statistical analysis for ecological studies.
Requires pre-registration on the Google Earth Engine platform (https://earthengine.google.com/).
}
\details{
Google Earth Engine main catalog
\itemize{
\item \link{get.ecors}	Get data to ecors
\item \link{stats.ecors}	Descritive statistics on ecors objects
\item \link{plot.ecors}	Plot ecors object
\item \link{download.ecors}	Download images
}
\cr
MapBiomas catalog
\itemize{
\item \link{get.mb.ecors}	Get MapBiomas data to ecors
\item \link{dist.mb.ecors}	Distance to Land Use class
\item \link{count.mb.ecors}	Count pixels of each Land Use class
\item \link{plot.focal.mb.ecors}	Plot mb.ecors object
\item \link{plot.mb.ecors}	Plot mb.ecors object
\item \link{download.mb.ecors}	Download MapBiomas images
}
}
\seealso{
\itemize{
\item \url{https://earthengine.google.com/}
\item \url{https://github.com/r-spatial/rgee/} or \link{rgee}
}
}
