% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecors-package.R
\docType{package}
\name{ecors}
\alias{ecors}
\title{ecors: Remote Sensing data acquisition and processing for ecology with Google Earth Engine}
\description{
Package to simplify the acquisition of data in Google Earth Engine and performs several steps of organization and processing typical of ecological studies, allowing pre-processing, visualization, data download and calculation of simple statistics commonly used in ecological studies. It aims to simplify the use of rgee and Google Earth Engine commands, performing internally several data manipulation steps. The function syntax follows a familiar style for R users who routinely perform statistical analysis for ecological studies. Requires pre-registration on the Google Earth Engine platform (https://earthengine.google.com/).
\cr
}
\details{
Google Earth Engine main catalog - remote sensing data
\itemize{
\item \link{get.ecors}	Get data to ecors
\item \link{stats.ecors}	Descritive statistics on ecors objects
\item \link{plot.ecors}	Plot ecors object
\item \link{download.ecors}	Download images
}
\cr
Land Use data
\itemize{
\item \link{get.lu.ecors}	Get Land Use data to ecors
\item \link{focal.lu.ecors} Applies a morphological reducer filter (focal) on Land Use images
\item \link{convert.lu.ecors} Re-assign land use classes by converting pixel values
\item \link{quantify.lu.ecors}	Quantify the pixels of each Land Use class
\item \link{dist.lu.ecors}	Distance to Land Use class
\item \link{plot.lu.ecors}	Plot mb.ecors object
\item \link{download.lu.ecors}	Download Land Use images
}

\cr
Misc
\itemize{
\item \link{wakeup.gee} If you receive an error message regarding the Google Earth Engine server being disconnected due to inactivity, use this command to re-establish the connection
}
}
\seealso{
\itemize{
\item \url{https://earthengine.google.com/}
\item \url{https://github.com/r-spatial/rgee/} or \link{rgee}
}
}
