% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ecors.R
\name{plot.ecors}
\alias{plot.ecors}
\title{Plot ecors object}
\usage{
\method{plot}{ecors}(
  x,
  ecors.type,
  visualization = "vis.bands",
  zoom = 10,
  defaults = T,
  legend = F,
  bands = NULL,
  pixel.min = NULL,
  pixel.max = NULL,
  image.gamma = NULL
)
}
\arguments{
\item{x}{ecors object (from get.ecors).}

\item{ecors.type}{type of images processed by get.ecors. Options are "original", "filtered", "mask", "filtered|mask", "filtered+mask", "composite". More information in detais section.}

\item{visualization}{choose preset options "vis.bands" (default visualization), "pan" (panchromatic band), "pan.sharpening" (panchromatic sharpening) or use "custom" to select bands.}

\item{zoom}{initial zoom value.}

\item{defaults}{if TRUE, uses default values of pixel.min, pixel.max, image.gamma and/or rescaling.}

\item{legend}{choose to show the legend or not.}

\item{bands}{select a single band to plot or three bands to plot RGB image.}

\item{pixel.min}{lower pixel value used to scale the plot.}

\item{pixel.max}{higher pixel value used to scale the plot.}

\item{image.gamma}{gamma correction.}
}
\value{

}
\description{
Plot Google Earth Engine images processed by get.ecors along with site and sample polygons.
}
\details{
Argument ecors.type selects which processing level of get.ecors will be ploted. Option "original" use all images available in the period without processing, "filtered" use only images approved in get.ecors quality control.
Option "mask" use same images as the previous one but with bad pixels masked, "composite" use compositions performed on the images of the previous option.
Options "filtered|mask" and "filtered+mask" show images from both collections. Option 2 and 3 show images from both collections, in the first the images are toggled with the slider and in the last option the images from both collections are stacked. in the first the images are toggled with the slider and in the last option the images from both collections are stacked
}
\examples{
#get a ecors class object
FAL.IBGE.JBB<-st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.retangles<-st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))
test.points<-st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))

d2020<-get.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles, buffer.points=500, buffer.plots=500,
    eval.area="site", projected=F, custom.crs=32723,
    collection="LANDSAT/LC08/C02/T1_L2", start=c("2020-01-01"), end=c("2021-01-01"),
    bands.eval="B3", bands.vis=T, indices=c("NDVI"), resolution=30,
    pOK=0.3, c.prob=NULL, c.dist=100, clouds.sentinel=NULL, cirrus.threshold=NULL, NIR.threshold=NULL, CDI.threshold=NULL, dmax.shadow=NULL,
    seasons=list(s1=c(11,12,1,2), s2=c(3,4), s3=c(5,6,7,8), s4=c(9,10)), sort.by="season", composite=NULL)

#ploting
plot.ecors(x=d2020, ecors.type="mask", visualization="vis.bands", defaults=T,
           legend=T, bands=NULL, pixel.min=NULL, pixel.max=NULL, image.gamma=NULL)

plot.ecors(x=d2020, ecors.type="mask", visualization="custom", defaults=F,
           legend=T, bands="NDVI", pixel.min=NULL, pixel.max=NULL, image.gamma=NULL)

}
