% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.ecors.R
\name{stats.ecors}
\alias{stats.ecors}
\title{Descritive statistics on ecors objects}
\usage{
stats.ecors(
  x,
  edge.pixels = "weighted",
  remove.samples = list(num.pixelOK = NULL, prop.pixelOK = 0.9),
  summarizing = "all",
  by.image.save = T,
  summary.save = T,
  stats = list(mean = T, median = T, sd = T),
  spreadsheet.folder = getwd()
)
}
\arguments{
\item{x}{ecors object (from get.ecors).}

\item{edge.pixels}{choose the treatment for pixels that coincide with the edges of the polygons: "weighted" uses the value of these pixels in the descriptive statistic weighted by the proportion of the area inside the polygon, "centroid" ignores pixels with the centroid outside the polygon.}

\item{remove.samples}{removes sample units that were misrepresented due to bad pixel removal. Minimum number of good pixels (num.pixelOK) and/or minimum proportion of good pixels (prop.pixelOK) can be selected.}

\item{summarizing}{selects whether the data will be summarized by year ("yearly") or for the entire evaluated period ("all"). The "yearly" option considers a year as every 12 months starting from the initial month of image collection.}

\item{by.image.save}{save a csv file with descriptive statistics for each image?}

\item{summary.save}{save a csv file with descriptive statistics summarizing all the images?}

\item{stats}{enables/disables the calculation of mean, median or standard deviation (sd).}

\item{spreadsheet.folder}{local folder to save csv files.}
}
\value{
List with result tables and metadata. If enabled, csv files will be saved with tables (two for each descriptive statistics): results.mean.samples.csv, summary.mean.samples.csv, results.median.samples.csv, summary.median.samples.csv, results.sd.samples.csv, summary.sd.samples.csv.
}
\description{
Performs descriptive statistics on study polygons at defined sampling periods.
}
\examples{
#get a ecors class object
FAL.IBGE.JBB<-sf::st_read(system.file("extdata/FAL.IBGE.JBB.gpkg", package="ecors"))
test.retangles<-sf::st_read(system.file("extdata/Plots_tests.gpkg", package="ecors"))
test.points<-sf::st_read(system.file("extdata/Points_tests.gpkg", package="ecors"))

d2020<-get.ecors(site=FAL.IBGE.JBB, points=test.points, plots=test.retangles, buffer.points=500, buffer.plots=500,
    eval.area="site", projected=F, custom.crs=32723,
    collection="LANDSAT/LC08/C02/T1_L2", start=c("2020-01-01"), end=c("2021-01-01"),
    bands.eval="SR_B3", bands.vis=T, indices=c("NDVI"), resolution=30,
    pOK=0.3, c.prob=NULL, c.dist=100, clouds.sentinel=NULL, cirrus.threshold=NULL, NIR.threshold=NULL, CDI.threshold=NULL, dmax.shadow=NULL,
    seasons=list(s1=c(11,12,1,2), s2=c(3,4), s3=c(5,6,7,8), s4=c(9,10)), sort.by="season", composite=NULL)

resp<-stats.ecors(x=d2020, edge.pixels="weighted", remove.samples=list(num.pixelOK=10,prop.pixelOK=0.8),
                  summarizing="all", by.image.save=T, summary.save=T,
                  stats=list(mean=T,median=F,sd=F),spreadsheet.folder=getwd() )
}
